backend/requirements.txt
# Tryloop — Python dependencies
# Versions will be pinned once initial development stabilizes.

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
redis>=5.0.0
stripe>=8.0.0
httpx>=0.27.0
emails>=0.6
python-dotenv>=1.0.0
